# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kartówka API",
    "description": "The Kartówka REST-like API.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/assets/{assetId}": {
      "patch": {
        "tags": [
          "Assets"
        ],
        "summary": "Update asset",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetDetailsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetDetailsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetDetailsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Assets"
        ],
        "summary": "Remove asset",
        "description": "This endpoint is idempotent.",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful no content result."
          }
        }
      }
    },
...
    "/api/packs/{packId}": {
      "get": {
        "tags": [
          "Packs"
        ],
        "summary": "Get pack by ID",
        "parameters": [
          {
            "name": "packId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "properties",
            "in": "query",
            "description": "Specifies what related entities should be pulled as well.",
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PackProperties"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pack object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Packs"
        ],
        "summary": "Update a pack",
        "parameters": [
          {
            "name": "packId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated pack",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Packs"
        ],
        "summary": "Remove a pack",
        "description": "This endpoint is idempotent.",
        "parameters": [
          {
            "name": "packId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content success result"
          }
        }
      }
    },
    "/api/packs": {
      "post": {
        "tags": [
          "Packs"
        ],
        "summary": "Create a new pack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pack object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pack"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/packs/{packId}/assets/{fileName}": {
      "post": {
        "tags": [
          "Packs"
        ],
        "summary": "Upload a new asset",
        "parameters": [
          {
            "name": "packId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Uploaded asset descriptor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pack not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Asset": {
        "required": [
          "blobUrl",
          "displayName",
          "systemName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "systemName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "size": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "assetType": {
            "$ref": "#/components/schemas/AssetType"
          },
          "blobUrl": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AssetType": {
        "enum": [
          "image",
          "music"
        ],
        "type": "string"
      },
      "CreatePackDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Pack": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/PackStatus"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            },
            "nullable": true
          },
          "questionsCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionsCategory"
            },
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "nullable": true
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Round"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PackProperties": {
        "enum": [
          "assets",
          "categories",
          "questions",
          "rounds"
        ],
        "type": "string"
      },
      "PackStatus": {
        "enum": [
          "draft",
          "active",
          "inactive"
        ],
        "type": "string"
      },
      "Question": {
        "required": [
          "answer"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "questionsCategoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "questionText": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "score": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "answer": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          },
          "contentType": {
            "$ref": "#/components/schemas/QuestionContentType"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "asset": {
            "$ref": "#/components/schemas/Asset"
          }
        },
        "additionalProperties": false
      },
      "QuestionContentType": {
        "enum": [
          "text",
          "image",
          "music"
        ],
        "type": "string"
      },
      "QuestionType": {
        "enum": [
          "default",
          "pass",
          "auction",
          "sponsored"
        ],
        "type": "string"
      },
      "QuestionsCategory": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roundId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "order": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Round": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "order": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAssetDetailsDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdatePackDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationResult": {
        "type": "object",
        "properties": {
          "memberNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put JWT token here",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}